[refactor] Split API REST Endpoints
